requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.2.14
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
